requirements.txt
django >3.2,<3.3
python-decouple
djangorestframework
markdown       
django-filter
django-filter
drf-yasg[validation]
djangorestframework_simplejwt
pytest
pytest-django
faker